LayoutTests/js/mozilla/strict/11.1.5-expected.txt

FAIL Function("'use strict'; ({x:1, x:1})") should throw an instance of SyntaxError
PASS !!Function("({x:1, x:1})") is true
PASS true === true
PASS !!Function("'use strict'; ({x:1, y:1})") is true
PASS !!Function("({x:1, y:1})") is true
PASS true === true
FAIL Function("'use strict'; ({x:1, y:1, x:1})") should throw an instance of SyntaxError
PASS !!Function("({x:1, y:1, x:1})") is true
PASS true === true
FAIL Function("'use strict'; ({x:1,   \"x\":1})") should throw an instance of SyntaxError
PASS !!Function("({x:1,   \"x\":1})") is true
PASS true === true
FAIL Function("'use strict'; ({\"x\":1, x:1})") should throw an instance of SyntaxError
PASS !!Function("({\"x\":1, x:1})") is true
PASS true === true
FAIL Function("'use strict'; ({\"x\":1, \"x\":1})") should throw an instance of SyntaxError
PASS !!Function("({\"x\":1, \"x\":1})") is true
PASS true === true
FAIL Function("'use strict'; ({1.5:1, 1.5:1})") should throw an instance of SyntaxError
PASS !!Function("({1.5:1, 1.5:1})") is true
PASS true === true
FAIL Function("'use strict'; ({1.5:1, 15e-1:1})") should throw an instance of SyntaxError
PASS !!Function("({1.5:1, 15e-1:1})") is true
PASS true === true
FAIL Function("'use strict'; ({6.02214179e23:1, 6.02214179e23:1})") should throw an instance of SyntaxError
PASS !!Function("({6.02214179e23:1, 6.02214179e23:1})") is true
PASS true === true
PASS !!Function("'use strict'; ({6.02214179e23:1, 3.1415926535:1})") is true
PASS !!Function("({6.02214179e23:1, 3.1415926535:1})") is true
PASS true === true
FAIL Function("'use strict'; ({ 1: 1, \"1\": 2 })") should throw an instance of SyntaxError
PASS !!Function("({ 1: 1, \"1\": 2 })") is true
PASS true === true
FAIL Function("'use strict'; ({ \"1\": 1, 1: 2 })") should throw an instance of SyntaxError
PASS !!Function("({ \"1\": 1, 1: 2 })") is true
PASS true === true
FAIL Function("'use strict'; ({ 2.5: 1, \"2.5\": 2 })") should throw an instance of SyntaxError
PASS !!Function("({ 2.5: 1, \"2.5\": 2 })") is true
PASS true === true
FAIL Function("'use strict'; ({ \"2.5\": 1, 2.5: 2 })") should throw an instance of SyntaxError
PASS !!Function("({ \"2.5\": 1, 2.5: 2 })") is true
PASS true === true
PASS !!Function("'use strict'; ({a:1, b:1, c:1, d:1, e:1, f:1, g:1, h:1, i:1, j:1, k:1, l:1, m:1, n:1, o:1, p:1, q:1, r:1, s:1, t:1, u:1, v:1, w:1, x:1, y:1, z:1})") is true
PASS !!Function("({a:1, b:1, c:1, d:1, e:1, f:1, g:1, h:1, i:1, j:1, k:1, l:1, m:1, n:1, o:1, p:1, q:1, r:1, s:1, t:1, u:1, v:1, w:1, x:1, y:1, z:1})") is true
PASS true === true
FAIL Function("'use strict'; ({a:1, b:1, c:1, d:1, e:1, f:1, g:1, h:1, i:1, j:1, k:1, l:1, m:1, n:1, o:1, p:1, q:1, r:1, s:1, t:1, u:1, v:1, w:1, x:1, y:1, a:1})") should throw an instance of SyntaxError
PASS !!Function("({a:1, b:1, c:1, d:1, e:1, f:1, g:1, h:1, i:1, j:1, k:1, l:1, m:1, n:1, o:1, p:1, q:1, r:1, s:1, t:1, u:1, v:1, w:1, x:1, y:1, a:1})") is true
PASS true === true
FAIL Function("'use strict'; ({get x() {}, x:1})") should throw an instance of SyntaxError
FAIL Function("({get x() {}, x:1})") should throw an instance of SyntaxError
PASS true === true
FAIL Function("'use strict'; ({x:1, get x() {}})") should throw an instance of SyntaxError
FAIL Function("({x:1, get x() {}})") should throw an instance of SyntaxError
PASS true === true
FAIL Function("'use strict'; ({set x(q) {}, x:1})") should throw an instance of SyntaxError
FAIL Function("({set x(q) {}, x:1})") should throw an instance of SyntaxError
PASS true === true
FAIL Function("'use strict'; ({x:1, set x(q) {}})") should throw an instance of SyntaxError
FAIL Function("({x:1, set x(q) {}})") should throw an instance of SyntaxError
PASS true === true
FAIL Function("'use strict'; ({1:1, set 1(q) {}})") should throw an instance of SyntaxError
FAIL Function("({1:1, set 1(q) {}})") should throw an instance of SyntaxError
PASS true === true
FAIL Function("'use strict'; ({set 1(q) {}, 1:1})") should throw an instance of SyntaxError
FAIL Function("({set 1(q) {}, 1:1})") should throw an instance of SyntaxError
PASS true === true
FAIL Function("'use strict'; ({\"1\":1, set 1(q) {}})") should throw an instance of SyntaxError
FAIL Function("({\"1\":1, set 1(q) {}})") should throw an instance of SyntaxError
PASS true === true
FAIL Function("'use strict'; ({set 1(q) {}, \"1\":1})") should throw an instance of SyntaxError
FAIL Function("({set 1(q) {}, \"1\":1})") should throw an instance of SyntaxError
PASS true === true
PASS !!Function("'use strict'; ({get x() {}, set x(q) {}})") is true
PASS !!Function("({get x() {}, set x(q) {}})") is true
PASS true === true
PASS !!Function("'use strict'; ({set x(q) {}, get x() {}})") is true
PASS !!Function("({set x(q) {}, get x() {}})") is true
PASS true === true
FAIL Function("'use strict'; ({get x() {}, set x(q) {}, x:1})") should throw an instance of SyntaxError
FAIL Function("({get x() {}, set x(q) {}, x:1})") should throw an instance of SyntaxError
PASS true === true
FAIL Function("'use strict'; ({set x(q) {}, get x() {}, x:1})") should throw an instance of SyntaxError
FAIL Function("({set x(q) {}, get x() {}, x:1})") should throw an instance of SyntaxError
PASS true === true
FAIL Function("'use strict'; ({get x() {}, get x() {}})") should throw an instance of SyntaxError
FAIL Function("({get x() {}, get x() {}})") should throw an instance of SyntaxError
PASS true === true
PASS Function("'use strict'; ({set x() {}, set x() {}})") threw exception of type SyntaxError.
PASS Function("({set x() {}, set x() {}})") threw exception of type SyntaxError.
PASS true === true
PASS !!Function("'use strict'; ({get x() {}, set x(q) {}, y:1})") is true
PASS !!Function("({get x() {}, set x(q) {}, y:1})") is true
PASS true === true
 PASSED!
PASS successfullyParsed is true

TEST COMPLETE
